Update style.css
 If we set font size in 'em' in body for 1em and 3em in '.title' - the title will be have one font size.
If we dont change font size in title (still 3 em), but change it in 'body' for 2em - the '.title' will change
his font size.
It shows us the dependence class selector 'title' from type selector 'body'.
The one selector inherits value from another.
Because one is parent selector, another is child selector. It depends which one of them is nested in other selector.